Update README.md
Uber Fare Prediction is a machine learning-based tool designed to estimate the cost of an Uber ride. It takes into account factors such as the pickup and dropoff locations, passenger count, car type, and the day of the week. The model adjusts the fare based on whether the ride is during a weekday or weekend, as well as the type of car chosen (e.g., Mini, XUV). This tool helps users quickly and accurately anticipate their ride costs, making trip planning more convenient and transparent.